requirements.txt
PySide6
pyinstaller
python-docx
python-barcode
pillow
lxml
